Clippy and Format, Circle Radius
Ran clippy fix and cargo fmt

Changed the radius of the circle from 4 -> 9 like the original
Change the colour of grass from green to dark green

Palette
Clippy got mad I had 0xff / 0xff
If that is the case, this might not be a great format for storing
palettes. It hurts the copy/pasteability of it.
I just wanted an easy way to copy/paste hex codes and not have to manually
convert them to 0.0-1.0 rgb values. So it was quite the failure in that
regard.