Health Bars and Deleting Towers
enemies.rs
move the shadow after the enemy.
On the web build, the shadow wasn't always following correctly. This aims
to fix it.

When building towers before, the background of the health bar would stick
around. Now it gets removed.

Gave sprites I wanted to access their own components so they are easier to
work with. It helps when multiple sprites are attached to the same entity.
Which may be a bad idea itself, but it's what I've done.

PileSprite to delete the pile sprite.
A gold pile has a GoldPile component
A tower preview has a GoldPile component
They both have sprites too.
But a tower preview's sprite gets swapped for the built tower sprite.
So they can't be destroyed quite the same. So only delete the PileSprite
entity.

TowerSprite
So only this sprite gets swapped, not every sprite attached to the hex.
The grass sprite was getting turned into a tower sprite. Then when the
tower was deleted, it looked like it stuck around, but it was just the
resprited grass.

TowerRemoveEvent
Used to just piggy-back on PileRemovedEvent
Now it's more clear what's happening.

When a pile attached to a tower preview gets full, it spawns a tower. This
is you paying for it. The full gold pile was just removed.
Now it is emptied to pay for the tower then it is removed by the same
process that removes piles when you press X on them. When you remove piles
this way, they dump their gold. So it's removed first so you don't get
infinite gold.