Gold, Enemy, Pile Bug
Maybe fixed a bug.
If an enemy is trying to steal a gold and you are trying to drop the gold
off at a GoldPile and they collide on the same frame, the gold will be
destroyed by one of them and then the other will try to use it.
Now the order is consistent. Storing gold happens before stealing. Maybe
this will fix it? I don't know when queries are made. At the start of
frame or just before the system runs?
At any rate, it's hard to test.