Gold Collision Bug
Tested the gold collision bug
Spawn an enemy and gold on top of a gold pile and see what happens

Gold was getting delted by gold pile and edited by enemy.grab
Moved gold collision together.
Now it obeys order of operations.
Priority to storing gold in the pile. If it gets stored, skip checking the
enemy collision and move on to the next gold in the query.

Press T for test
Probably should have removed this so I can release this