Growing Hexes
You start on a smaller area. But it grows over time. It grows a new side
every 10s. This would mimic a clock's minute hand spinning around.

Fixed bugs
Towers were eating too much gold if they collided on the same frame.
Health bar was sticking around if the tower was build the same frame it
was clicked to make a preview.