Tutorial
Made a tutorial on the side of the screen.
Press tab to dismiss it
Changed the bg colour so you can see it from the main menu

Made the game easier
Spawn Gold: 6 -> 11
Enemy Group: 10 -> 5
Makes it so you aren't sunk if you can't figure out the game before
enemies spawn.

Bug Fix
When building a tower, it would suck up all gold colliding with it that frame.