Fancy Transition
Made a fancy transition. Press T to toggle it.
Just made it to see if I could. It isn't really used for anything.
There's a weird bug with the spinning, but it looks cool so i'm leaving
it.

Art
Made a different grass tile. It's not much better, I was just trying to
see if I could tile hexes with pixel art. It works decently. I think I'd
have to do the positioning calculations based on individual pixels (to the
left 15px, up 13px, etc.) to get it to really tile well.
Changed the size of the hexes to match.
Having a hex with an outline is nice because then you can see the grid
without hovering over it to see the relationship between the towers.
(An earlier version of the grass had an outline, not this current one.)