Info Panel
When you right click, make a little info box.
It tells you what the hex where you clicked is.
Hex, Gold Pile or Tower
Default is "This is a Hex" so if you click nothing, it will say it is a
hex.
It's more of a proof-of-concept
It has no logic to not cover up what you are looking at. If you click on
the right edge of the screen, it will appear off screen.

Bug
I tied it to the selection. When you got a new selection, it would clear
it.
Then I learned I could use Interaction like buttons do.
But now it's still coupled to the selection.
If I remove the selection code in remove_old_panel, the panel is removed
immediately, even if right under the mouse.
Why is this?

Future
I want it to delete after a second of not hovering, not immediately. This
might fix the bug in a roundabout way.