Bomb Bullet
Made a bomb bullet. Fires in an arc to its target.
I doesn't no anything yet. Just got the movement working.

Flies up into the air in an arc before coming down on target.
Has constant lifetime. The arc always takes 1.0s regardless of distance.

The idea is to have it arc and when it lands, spawn an explosion. An
explosion would just be a large circle collider.
While flying in the arc, the bomb bullet has no collision (it is above the
enemies' head).

Aiming
Doesn't attempt to aim. Just shoots at where an enemy is when it comes
into range.
This is the same as normal bullets. Currently, whenever a tower shoots a
bullet, it also shoots a bomb bullet.

This may not be the best way to do this.
There are a lot of parameters. Like playing with the intial arc
direction's length and the blending easing functions used.

Might be better to have an implementation that has constant arc height and
can adjust hangtime. Then the length of the arc changes to shoot farther
or nearer.
Constant from spawn or target? Probably target so it looks normal shooting
up.