Refactoring
Moved gold.setup to the top

remove_pile
better destructuring. Gave stuff names instead of child_query.0 or .2,
etc.

pile_input
removed some testing buttons
Press T to spawn some stuff for testing

ShootType
Might be a better way to make multiple types of towers
Still a little awkward. The Tower knows its shoot type, but you spawn a
TowerPreview, not a Tower. So there's a system to update it when the Tower
is made.
But maybe I could add Tower and TowerPreview at the start. And then Tower
Queries would include Without<TowerPreview>
Or the TowerPreview could pass the shoottype along?

remove_tower
Cleaned up stuff.
Had a weird counter that was just some or checks
Coupled the refund logic. It was an if, then a loop. Put the loop in the
if to be more clear that that was the only place that variable was used.

tower_shoot
Each shootType gets its own logic.
Bullet cares about direction
Arc cares about position and prediction