HexCollection, Gold Gen Radius
HexCollection
Store hexes in a hashmap when they are made.
map<HexCoords, Entity>
Use a hexcoord to get the hex at that location if it exists

Now I could replace everywhere where I was looping over all hexes trying
to match the event coords with each of them with just looking up the hex
using the coords.

This allowed me to change most queries grabbing the hex to just With<Hex>

generate_gold()
Now works with variable radii.
Towers produce gold in the 6 tiles around them.
Bomb Towers produce gold in the 6 neighbour tiles an the 12 tiles on the
next ring.