Bombs and Rapier2D
Bomb towers now work.
When the arc ends, it makes an explosion. The explosion is dangerous for
half of its lifetime. It's dangerous for 0.15s. After that, enemies can
enter without dying. This would match up with a potential animation of the
smoke dissappating before the explosion is destroyed.

Added rapier for collision between the bomb and enemies
It allows me to do circle collisions.
Other collisions aren't converted yet.

Updated to bevy 8.1
The rapier version I used uses 8.1
It probably doesn't make a difference

Changed the colour of health bar/ progress bars to gold. They measure gold
in the pile so this is more apt.